+++ database/migrations/2021_03_13_050756_create_category_masters_table.php
@@ -21,7 +21,7 @@ class CreateCategoryMastersTable extends Migration
             $table->foreign('module_type_id')->references('id')->on('module_types')->onDelete('cascade');
             $table->string('category_master_id', 50)->nullable();
             //$table->foreign('category_master_id')->references('id')->on('category_masters')->onDelete('cascade');
-            $table->float('vat', 8,2)->default(0);
+            $table->double('vat', 12,2)->default(0);
 
             $table->string('title');
             $table->string('slug');
